force selection of problem type; hardcoded layout
We now only generate anything after you pick what kind of
problem you want.  This avoids annoyance with deciding
what the default will be and just generally simplifies
stuff.

Also hard code the hell out of the layout and do everything
inches.  This is not something I like, but nothing else I've
tried can seem to convince Chrome not to cut off the top and
left when printing.  So, runs good (as long as you are using
8.5 x 11 inch paper :P)