added multi-digit multiplication
Added multi-digit multiplication but it's a little intense
because it's the same number of problems.  Also changed the
structure of the things you can select so that I can add a
default number of problems and have it be lower for (eg)
multi-digit multiplication.  This would mean it would
(probably?) have to change the table or whatever it's writing
to, which was not the goal here but whatevs.  Probably best.