choose/set problems off search params
Make the selector reload the page with appropriate search params.

Make it use the search params instead of the selector update to
fill in the problems.

This is better for making links to particular pages, and also
means you can use the back button (though you won't necessarily
get the same problems since it's all client side/dynamic).